Implement using callbacks, promises and async/await
